backend/requirements.txt
yfinance<0.2.32
pandas>=2.0.0
numpy>=1.24.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
groq>=0.4.0
requests>=2.31.0
python-dotenv>=1.0.0
kiteconnect>=4.0.0